+++ src/Indice.Features.Messages.Core/Data/Models/DbMessageEvent.cs
@@ -7,6 +7,8 @@ public class DbMessageEvent
     public Guid Id { get; set; } = Guid.NewGuid();
     /// <summary>The unique identifier of the associated campaign.</summary>
     public Guid CampaignId { get; set; }
+    /// <summary>The title of the message.</summary>
+    public string Title { get; set; } = string.Empty;
     /// <summary>The unique identifier of the associated contact.</summary>
     public Guid ContactId { get; set; }
     /// <summary>The unique identifier of the message.</summary>
@@ -18,7 +20,9 @@ public class DbMessageEvent
     /// <summary>The recipient depending on the channel.</summary>
     /// <remarks>Phone number for sms, email for email or recipient Id for other channels.</remarks>
     public string Recipient { get; set; } = string.Empty;
-    
+    /// <summary>Indicates whether the event was successful.</summary>
+    public bool Success { get; set; } = false;
+
     /// <summary>The date and time when the event occurred.</summary>   
     public DateTimeOffset CreatedOn { get; set; } = DateTimeOffset.UtcNow;
 }