+++ src/Indice.Features.Messages.Core/Events/SendPushNotificationEvent.cs
@@ -60,13 +60,16 @@ public class SendPushNotificationEvent
     /// Converts the current <see cref="SendPushNotificationEvent"/> to a <see cref="MessageEvent"/> with the specified type.
     /// </summary>
     /// <param name="type">the type of the event</param>
+    /// <param name="succeded">indicate the event status</param>
     /// <returns></returns>
-    public MessageEvent ToMessageEvent(string type) => new MessageEvent {
+    public MessageEvent ToMessageEvent(string type, bool succeded) => new MessageEvent {
         CampaignId = CampaignId,
         ContactId = ContactId,
         MessageId = MessageId,
         Type = type,
         Channel = MessageChannelKind.PushNotification.ToString(),
         Recipient = RecipientId!,
+        Title = Title ?? "",
+        Success = succeded
     };
 }