+++ src/Indice.Features.Messages.Core/Events/SendSmsEvent.cs
@@ -52,13 +52,16 @@ public class SendSmsEvent
     /// Converts the current <see cref="SendSmsEvent"/> to a <see cref="MessageEvent"/> with the specified type.
     /// </summary>
     /// <param name="type">the type of the event</param>
+    /// <param name="succeded">indicate the event status</param>
     /// <returns></returns>
-    public MessageEvent ToMessageEvent(string type) => new MessageEvent {
+    public MessageEvent ToMessageEvent(string type, bool succeded) => new MessageEvent {
         CampaignId = CampaignId,
         ContactId = ContactId,
         MessageId = MessageId,
         Type = type,
         Channel = MessageChannelKind.SMS.ToString(),
-        Recipient = RecipientPhoneNumber!
+        Recipient = RecipientPhoneNumber!,
+        Title = Title ?? "",
+        Success = succeded
     };
 }