+++ src/Indice.Features.Messages.App/src/app/app.module.ts
@@ -93,6 +93,7 @@ import { MediaSettingEditComponent } from './features/settings/media/edit/media-
 import { CodeEditorModule } from '@acrodata/code-editor';
 import { NgProgressbar } from 'ngx-progressbar';
 import { progressInterceptor, NgProgressHttp } from 'ngx-progressbar/http';
+import { MessageEventsComponent } from './features/events/message-events.component';
 registerLocaleData(localeGreek);
 
 const providers: Provider[] = [
@@ -200,7 +201,8 @@ if (app.settings.tenantId) {
     MediaSettingsComponent,
     MediaSettingEditComponent,
     DoughnutChartComponent,
-    LineChartComponent
+    LineChartComponent,
+    MessageEventsComponent
   ],
   imports: [
     AppRoutingModule,